refactor(filterpb): adopt .v1 FQN package
Move filter.proto and its Go companions into a v1/ subdirectory and rename
the package from filterpb to common.filterpb.v1, bringing the shared filter
types in line with the .v1 FQN convention. The generated Go package name
stays filterpb via the go_package alias. Switches the meson build to the
collision-safe -I root_dir strategy so the file registers under its full
repo-relative path. Updates importing protos, Go importers, and Rust
extern_path/include_proto wiring, including the route-mpls CLI's local
filterpb include (kept local for its orphan-rule trait impl).
