+++ common/filterpb/meson.build
@@ -2,10 +2,9 @@ filter_proto_dir = meson.current_source_dir()
 root_dir = meson.project_source_root()
 
 filter_proto_files = [
-    join_paths(filter_proto_dir, 'filter.proto'),
+    join_paths(filter_proto_dir, 'v1', 'filter.proto'),
 ]
 
-# Generate protobuf files
 filter_protoc_gen = custom_target(
     'filter-protoc',
     output: [
@@ -15,10 +14,9 @@ filter_protoc_gen = custom_target(
     input: filter_proto_files,
     command: [
         protoc,
-        '-I', filter_proto_dir,
         '-I', root_dir,
-        '--go_out=paths=source_relative:' + filter_proto_dir,
-        '--go-grpc_out=paths=source_relative:' + filter_proto_dir,
+        '--go_out=paths=source_relative:' + root_dir,
+        '--go-grpc_out=paths=source_relative:' + root_dir,
         '@INPUT@',
     ],
     build_by_default: true,
