+++ common/rust/filterpb/src/lib.rs
@@ -1,6 +1,6 @@
 #[allow(clippy::all, non_snake_case)]
 pub mod pb {
-    tonic::include_proto!("filterpb");
+    tonic::include_proto!("common.filterpb.v1");
 }
 
 pub mod network;
