+++ modules/acl/controlplane/aclpb/v1/acl.proto
@@ -2,7 +2,7 @@ syntax = "proto3";
 
 package modules.acl.controlplane.aclpb.v1;
 
-import "common/filterpb/filter.proto";
+import "common/filterpb/v1/filter.proto";
 import "common/commonpb/v1/metric.proto";
 
 option go_package = "github.com/yanet-platform/yanet2/modules/acl/controlplane/aclpb/v1;aclpb";
@@ -41,14 +41,14 @@ message Rule {
   repeated Action actions = 1;
   string counter = 2;
 
-  repeated filterpb.Device devices = 3;
-  repeated filterpb.VlanRange vlan_ranges = 4;
-  repeated filterpb.IPNet srcs = 5;
-  repeated filterpb.IPNet dsts = 6;
-  repeated filterpb.ProtoRange proto_ranges = 7;
-  repeated filterpb.PortRange src_port_ranges = 8;
-  repeated filterpb.PortRange dst_port_ranges = 9;
-  filterpb.Fragment fragment = 10;
+  repeated common.filterpb.v1.Device devices = 3;
+  repeated common.filterpb.v1.VlanRange vlan_ranges = 4;
+  repeated common.filterpb.v1.IPNet srcs = 5;
+  repeated common.filterpb.v1.IPNet dsts = 6;
+  repeated common.filterpb.v1.ProtoRange proto_ranges = 7;
+  repeated common.filterpb.v1.PortRange src_port_ranges = 8;
+  repeated common.filterpb.v1.PortRange dst_port_ranges = 9;
+  common.filterpb.v1.Fragment fragment = 10;
 }
 
 message Action { ActionKind kind = 1; }
