+++ modules/forward/cli/build.rs
@@ -6,7 +6,7 @@ pub fn main() -> Result<(), Box<dyn Error>> {
     tonic_build::configure()
         .emit_rerun_if_changed(false)
         .build_server(false)
-        .extern_path(".filterpb", "::filterpb::pb")
+        .extern_path(".common.filterpb.v1", "::filterpb::pb")
         .message_attribute(".", "#[derive(Serialize)]")
         .compile_protos(&["forwardpb/v1/forward.proto"], &["../../..", "../controlplane"])?;
