+++ modules/route-mpls/cli/build.rs
@@ -1,7 +1,7 @@
 use core::error::Error;
 
 pub fn main() -> Result<(), Box<dyn Error>> {
-    println!("cargo:rerun-if-changed=../../../common/filterpb/filter.proto");
+    println!("cargo:rerun-if-changed=../../../common/filterpb/v1/filter.proto");
     println!("cargo:rerun-if-changed=../../../common/commonpb/ipaddr.proto");
     println!("cargo:rerun-if-changed=../controlplane/routemplspb/v1/routempls.proto");
 
@@ -9,7 +9,7 @@ pub fn main() -> Result<(), Box<dyn Error>> {
         .emit_rerun_if_changed(false)
         .build_server(false)
         .extern_path(".common.commonpb.v1", "::commonpb::pb")
-        .extern_path(".filterpb", "crate::filterpb")
+        .extern_path(".common.filterpb.v1", "crate::filterpb")
         .message_attribute(".", "#[derive(Serialize)]")
         .enum_attribute(".", "#[derive(serde::Serialize)]")
         .compile_protos(&["routemplspb/v1/routempls.proto"], &["../../..", "../controlplane"])?;
@@ -18,7 +18,7 @@ pub fn main() -> Result<(), Box<dyn Error>> {
         .emit_rerun_if_changed(false)
         .build_server(false)
         .message_attribute(".", "#[derive(Serialize)]")
-        .compile_protos(&["common/filterpb/filter.proto"], &["../../.."])?;
+        .compile_protos(&["common/filterpb/v1/filter.proto"], &["../../.."])?;
 
     Ok(())
 }
